Comment addressing, refactor
