+++ ledger/notary-plugin/src/main/kotlin/net/corda/v5/ledger/notary/plugin/api/PluggableNotaryClientFlow.kt
@@ -1,4 +1,4 @@
-package net.corda.v5.ledger.notary.pluggable
+package net.corda.v5.ledger.notary.plugin.api
 
 import net.corda.v5.application.crypto.DigitalSignatureAndMetadata
 import net.corda.v5.application.flows.SubFlow
@@ -11,14 +11,14 @@ import net.corda.v5.application.flows.SubFlow
  * If the client flow doesn't implement this interface the plugin selection will not work.
  * To learn more about plugin selection and types see [PluggableNotaryType].
  *
- * Implementations must specify the [net.corda.v5.application.flows.InitiatingFlow] annotation.
- * Further, the [call] method must specify the [net.corda.v5.base.annotations.Suspendable] annotation.
+ * Implementations must specify the [InitiatingFlow][net.corda.v5.application.flows.InitiatingFlow] annotation.
+ * Further, the [call] method must specify the [Suspendable][net.corda.v5.base.annotations.Suspendable] annotation.
  *
  * The server side will not have an interface like the client, it will only implement the
- * [net.corda.v5.application.flows.ResponderFlow] interface.
+ * [ResponderFlow][net.corda.v5.application.flows.ResponderFlow] interface.
  *
  * For an example client and server implementation please refer to the non-validating notary plugin under
  * the `notary-plugins/notary-plugin-non-validating` module in the `corda-runtime-os` repository,
- * or the quick starter guide.
+ * or the quick start guide.
  */
 interface PluggableNotaryClientFlow : SubFlow<List<DigitalSignatureAndMetadata>>
