+++ ledger/notary-plugin/src/main/kotlin/net/corda/v5/ledger/notary/plugin/api/PluggableNotaryClientFlowProvider.kt
@@ -1,4 +1,4 @@
-package net.corda.v5.ledger.notary.pluggable
+package net.corda.v5.ledger.notary.plugin.api
 
 import net.corda.v5.ledger.common.Party
 import net.corda.v5.ledger.utxo.transaction.UtxoSignedTransaction
@@ -9,9 +9,6 @@ import net.corda.v5.ledger.utxo.transaction.UtxoSignedTransaction
  *
  * Implementers of this interface must have the [PluggableNotaryType] annotation,
  * otherwise an exception will be thrown on startup.
- *
- * The client-side of the plugin must have the same constructor parameters as the
- * [create] function (notary, stx, context).
  */
 interface PluggableNotaryClientFlowProvider {
     fun create(notary: Party, stx: UtxoSignedTransaction): PluggableNotaryClientFlow
