+++ ledger/notary-plugin/src/main/kotlin/net/corda/v5/ledger/notary/plugin/core/NotarisationResponse.kt
@@ -1,11 +1,11 @@
-package net.corda.v5.ledger.notary.pluggable
+package net.corda.v5.ledger.notary.plugin.core
 
 import net.corda.v5.application.crypto.DigitalSignatureAndMetadata
 import net.corda.v5.application.uniqueness.model.UniquenessCheckError
 
 /**
  * An interface that represents the payload returned by the notary server flow to the client flow. Contains the
- * signatures from the server side and the error, in case there's any.
+ * signatures from the server side and the error, if applicable.
  */
 interface NotarisationResponse {
     val signatures: List<DigitalSignatureAndMetadata>
